refactor: 불필요한 코드 삭제, int 형으로 정의한 금액 변수 long 형으로 수정

+++ src/Guest.java
@@ -3,11 +3,11 @@ public class Guest {
 
     private String name;
     private String phone;
-    private int cash;
+    private long cash;
     private boolean isNew;
 
 
-    public Guest(String name, String phone, int cash) {
+    public Guest(String name, String phone, long cash) {
         this.name = name;
         this.phone = phone;
         this.cash = cash;
@@ -22,11 +22,11 @@ public class Guest {
         return phone;
     }
 
-    public int getCash() {
+    public long getCash() {
         return cash;
     }
 
-    public void setCash(int cash) {
+    public void setCash(long cash) {
         this.cash = cash;
     }
